+++ .prettierrc.js
@@ -1,5 +1,5 @@
 module.exports = {
-	printWidth: 80,
-	useTabs: true,
-	tabWidth: 2
+  printWidth: 80,
+  singleQuote: true,
+  trailingComma: 'es5',
 };