+++ examples/dev/dist.js
@@ -2,4 +2,4 @@
 require('../../util/change_cwd')(__dirname + '/demo');
 
 // require('../../util/exec')('lila dist test/index -l senn -o');
-require('../../util/exec')('lila dist test-6/index');
+require('../../util/exec')('lila dist test-7');