+++ CHANGELOG.md
@@ -20,9 +20,9 @@
 
 ## 0.4.0 [2018-07-26]
 
-- Remove `archive,arc,clean,revert` all server-side commands. 
+- Remove `archive,arc,clean,revert` all server-side commands.
 - Remove `ana` command.
-- Fix bug of `eslint,stylelint,prettier` supporting `test/index,test-2/*` multi mode module. 
+- Fix bug of `eslint,stylelint,prettier` supporting `test/index,test-2/*` multi mode module.
 
 ## 0.3.2 [2018-07-10]